Move router definition to index
